avoid float->int conversion.
[git-p4: depot-paths = "//depot/qt/main/": change = 159306]

+++ src/gui/painting/qpaintengine_win_p.h
@@ -234,7 +234,7 @@ public:
 
     Qt::PenStyle penStyle;
     Qt::BrushStyle brushStyle;
-    int pWidth;
+    float pWidth;
     COLORREF pColor;
     COLORREF bColor;
     QBrush brush;
